folders structure refactor
