+++ src/threadx/SennaTech/Core/Inc/car/car.h
@@ -8,7 +8,6 @@
 #define PWM_MAX_RAW_VALUE ((1 << PWM_RESOLUTION) - 1)  
 #define PWM_FREQ_SERVO_HZ 0x79
 #define PWM_FREQ_MOTOR_HZ 0x05
-#define PWM_WAVELENGTH_50HZ (1.0f / PWM_FREQ_SERVO_HZ)
 #define PWM_STEERING_CHANNEL 0
 #define PWM_THROTTLE_CHANNEL_LEFT_MOTOR_IN_1 5
 #define PWM_THROTTLE_CHANNEL_LEFT_MOTOR_IN_2 6
@@ -21,10 +20,8 @@
 #define SERVO_RAW_MIN_MECH 205
 #define SERVO_RAW_MAX_MECH 410
 
-/* Safe operating range for MG996R (reduces stall risk) */
-/* Limit to ~80% of full range to prevent mechanical hard-stops */
-#define SERVO_RAW_MIN 230   /* ~85% from center towards left */
-#define SERVO_RAW_MAX 385   /* ~85% from center towards right */
+#define SERVO_RAW_MIN 215
+#define SERVO_RAW_MAX 400
 
 typedef struct
 {
