+++ src/threadx/SennaTech/Core/Inc/i2c/i2c_hal.h
