+++ src/threadx/SennaTech/Core/Inc/pid/pid.h
@@ -2,6 +2,7 @@
 #define PID_H
 
 #include <stdint.h>
+#include <utils.h>
 
 typedef struct
 {
@@ -19,9 +20,15 @@ float pid_update(pid_t *pid, float target, float current, float dt);
 void pid_reset(pid_t *pid);
 void pid_set_integral_limit(pid_t *pid, float limit);
 void pid_set_output_limit(pid_t *pid, float limit);
-float clamp_symmetric(float value, float limit);
 
-#define THROTTLE_PERIOD_TICKS 2
+//pid utils
+float absf_local(float value);
+float clamp_symmetric(float value, float limit);
+float get_throttle_current_normalized(void);
+float pidThrottleCalculation(pid_t *throttle_pid, float throttle_target, float dt);
+float get_delta_time_seconds(ULONG *last_tick);
+
+#define THROTTLE_PERIOD_TICKS 5
 
 // Real measured vehicle speed from sensor, global/shared
 extern float speed_kmh;
@@ -29,7 +36,6 @@ extern float speed_kmh;
 #endif
 
 /*
-
 Proportional gain -> reacts to current error
 Integral gain -> reacts to accumulated error over time
 Derivative gain -> reacts to the rate of change of error
