+++ src/threadx/SennaTech/Core/Inc/utils/utils.h
