+++ src/threadx/SennaTech/Core/Src/can/can_manager.c
