+++ src/threadx/SennaTech/Core/Src/heartbeat.c
@@ -7,12 +7,12 @@ void heartbeat_thread_entry(ULONG thread_input)
     hb_frame.dlc = 1;
     hb_frame.data[0] = 0xAA;
 
-    //log_debug("HEARTBEAT THREAD STARTED");
+    uart_send("HEARTBEAT THREAD STARTED\r\n");
 
     while(1)
     {
         tx_queue_send(&g_tx_data_queue, &hb_frame, TX_NO_WAIT);
 
-        tx_thread_sleep(100);
+        tx_thread_sleep(500);
     }
 }
