+++ src/threadx/SennaTech/Core/Src/i2c/i2c_hal.c
